+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [0.7.1] 2025-02-03
+
+### Added
+
+- set input and output port based on parameter values
+- sample entities to workflow execution report
 
 ## [0.7.0] 2025-01-06